refactor: (#18) useSocket shared 레이어로 분리
